renamed parent.ftl to parent-home.ftl; added security for parent-home page; added redirect to parent-home page after logging; added ParentController class;
