+++ src/main/java/com/inva/hipstertest/config/SecurityConfiguration.java
@@ -102,6 +102,7 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
             .antMatchers("/freemarker/logout").authenticated()
             .antMatchers("/freemarker/teacher-mgmt/**").hasAuthority(AuthoritiesConstants.HEAD_TEACHER)
             .antMatchers("/freemarker/pupil-home").hasAuthority(AuthoritiesConstants.PUPIL)
+            .antMatchers("/freemarker/parent-home").hasAuthority(AuthoritiesConstants.PARENT)
             .antMatchers("/freemarker/freemarkertest/**").permitAll() // change
 
             .antMatchers("/api/register").permitAll()
