+++ src/main/java/com/inva/hipstertest/freemarker/controllers/UserJWTFreemarkerController.java
@@ -69,6 +69,8 @@ public class UserJWTFreemarkerController {
                 return "redirect:pupil-home";
             } else if (authorities.contains(new SimpleGrantedAuthority("ROLE_HEAD_TEACHER"))) {
                 return "redirect:teacher-mgmt/teacher-mgmt";
+            } else if (authorities.contains(new SimpleGrantedAuthority("ROLE_PARENT"))) {
+                return "redirect:parent-home";
             } else {
                 return "redirect:freemarkertest";
             }
